backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.12
pypdf==5.1.0
faiss-cpu
numpy==2.1.3
google-genai==1.40.0
python-dotenv==1.0.1
PyJWT==2.10.1
requests==2.32.3
httpx==0.28.1
beautifulsoup4==4.12.3
supabase==2.18.1
boto3==1.40.54
